# Don't keep box muller transform state between kernel launches