… src/genn/backends/cuda/backend.cc
… src/genn/backends/hip/backend.cc
… src/genn/genn/code_generator/backendCUDAHIP.cc
… src/genn/genn/code_generator/backendSIMT.cc