This builds on the  new hybrid HIP/CUDA backend from #647 so review that first!

As described in #648, beyond its (arguably excessive) 192 bit of state, the curand XORWOW RNG used to provide randomness for neuron and custom connectivity updates also stores 160 bits of [box muller transform](https://en.wikipedia.org/wiki/Box%E2%80%93Muller_transform) state in the ``curandState`` struct (BM draws two numbers and produces two normally distributed values so this state is used to cache one of those results for subsequent calls to ``curand_normal``). 

In this PR, when using CUDA and HIP backends, we create our own ``XORWowStateInternal`` struct in definitions.h without the BM state and store this in memory. At the start of the neuron and custom connectivity update kernels, we copy the fields from the  ``XORWowStateInternal`` struct into a local ``curandState`` and, at the end, we copy them back.

Excitingly, because we are very memory bandwidth bound, this makes the neuron kernel on the cortical microcircuit model about 60% faster. On my A5000 (running for 1 second):

|                                | Before[seconds] | After [seconds] |
|--------------------- |-------------------|------------------|
|neuron update        | 0.18                    | 0.11                  |
|presynaptic update | 0.23                    | 0.24                  |